feat(workflows): add stage/prod build workflows (#637)
Replaces the `docker-build` workflow with two workflows
`prod-build` and `stage-build`, which both use the reusable
`_build` workflow to build and push a Docker image.

Previously, `docker-build` pushed to the GitHub Container Registry.
Now `_build` pushes to the Google Cloud Artifact Registry.

The workflows don't deploy directly, but the pushed images implicitly
trigger deployment to stage (for commits) and/or prod (for tags).